requirements.txt
docker
datasets
langchain
langchain-core
langchain-community
langgraph
langsmith
langchain-openai==0.1.20
python-dotenv
gitpython
flask
psycopg[binary]>=3.1.19
markupsafe==2.0.1
directory_tree
pylint
chromadb
<<<<<<< HEAD
#pysqlite3
git+https://github.com/princeton-nlp/SWE-bench.git#egg=swebench
=======
pysqlite3
git+https://github.com/princeton-nlp/SWE-bench.git#egg=swebench
langchain-huggingface
langchain-text-splitters
>>>>>>> 1f9cce2cbffcefed8ab95beb7f121572cb6b3459
